requirements.txt
torch>=1.8.0
torchvision>=0.9.0
pyyaml
matplotlib
tqdm
numpy
pycocotools
albumentations
